refactor: split marshal/unmarshal.go into three focused files
Extract JSON marshaling into unmarshal_json.go and variant reconstruction
into variant_reconstruct.go; split matching tests into dedicated test files.
