+++ marshal/unmarshal_test.go
@@ -143,53 +143,6 @@ func TestUnmarshal_PanicZeroValue(t *testing.T) {
 	require.Contains(t, err.Error(), "dstInterface must be a non-nil pointer")
 }
 
-func TestConvertToJSONFriendly_OldListFormat(t *testing.T) {
-	// Struct with single exported "Array" field of slice type simulates old list format
-	type OldList struct {
-		Array []int32 `json:"array"`
-	}
-
-	sh, err := schema.NewSchemaHandlerFromStruct(new(struct {
-		Name string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-	}))
-	require.NoError(t, err)
-
-	input := OldList{Array: []int32{1, 2, 3}}
-	result, err := ConvertToJSONFriendly(input, sh)
-	require.NoError(t, err)
-
-	// Should unwrap to a slice, not a map
-	slice, ok := result.([]any)
-	require.True(t, ok)
-	require.Equal(t, 3, len(slice))
-	require.Equal(t, int32(1), slice[0])
-	require.Equal(t, int32(2), slice[1])
-	require.Equal(t, int32(3), slice[2])
-}
-
-func TestConvertToJSONFriendly_UnexportedFields(t *testing.T) {
-	type MixedStruct struct {
-		Public  string `json:"public"`
-		private string //nolint:unused // intentionally unexported for test
-	}
-
-	sh, err := schema.NewSchemaHandlerFromStruct(new(struct {
-		Name string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-	}))
-	require.NoError(t, err)
-
-	input := MixedStruct{Public: "visible"}
-	result, err := ConvertToJSONFriendly(input, sh)
-	require.NoError(t, err)
-
-	m, ok := result.(map[string]any)
-	require.True(t, ok)
-	require.Contains(t, m, "public")
-	require.Equal(t, "visible", m["public"])
-	// Unexported field should not be in the result
-	require.NotContains(t, m, "private")
-}
-
 func TestUnmarshalStringToByteArray(t *testing.T) {
 	// Marshal stores string as BYTE_ARRAY, verify Unmarshal round-trip is consistent.
 	type ByteStruct struct {
@@ -214,497 +167,6 @@ func TestUnmarshalStringToByteArray(t *testing.T) {
 	require.Len(t, dst, 2)
 	require.Equal(t, "hello", dst[0].Data)
 	require.Equal(t, "world", dst[1].Data)
-}
-
-func TestConvertToJSONFriendly_Combined(t *testing.T) {
-	type DecimalStruct struct {
-		Decimal1 int32  `parquet:"name=decimal1, type=INT32, convertedtype=DECIMAL, scale=2, precision=9"`
-		Decimal2 int64  `parquet:"name=decimal2, type=INT64, convertedtype=DECIMAL, scale=3, precision=18"`
-		Decimal3 string `parquet:"name=decimal3, type=BYTE_ARRAY, convertedtype=DECIMAL, scale=4, precision=20"`
-		Decimal4 string `parquet:"name=decimal4, type=FIXED_LEN_BYTE_ARRAY, convertedtype=DECIMAL, scale=2, precision=10, length=12"`
-		Decimal5 int32  `parquet:"name=decimal5, type=INT32, logicaltype=DECIMAL, logicaltype.precision=9, logicaltype.scale=2"`
-		Name     string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-		Age      int32  `parquet:"name=age, type=INT32"`
-	}
-	decimalSchemaHandler, err := schema.NewSchemaHandlerFromStruct(new(DecimalStruct))
-	require.NoError(t, err)
-
-	type NestedStruct struct {
-		Value int32 `parquet:"name=value, type=INT32, convertedtype=DECIMAL, scale=2, precision=5"`
-	}
-	type ContainerStruct struct {
-		Nested   NestedStruct     `parquet:"name=nested"`
-		SliceVal []int32          `parquet:"name=slice_val, type=LIST, valuetype=INT32"`
-		MapVal   map[string]int32 `parquet:"name=map_val, type=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32"`
-		PtrVal   *int32           `parquet:"name=ptr_val, type=INT32, convertedtype=DECIMAL, scale=1, precision=4"`
-	}
-	containerSchemaHandler, err := schema.NewSchemaHandlerFromStruct(new(ContainerStruct))
-	require.NoError(t, err)
-
-	type SimpleStruct struct {
-		Name string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-	}
-	simpleSchemaHandler, err := schema.NewSchemaHandlerFromStruct(new(SimpleStruct))
-	require.NoError(t, err)
-
-	type NestedDecimalStruct struct {
-		MapWithDecimals  map[string]int32 `parquet:"name=map_with_decimals, type=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32, valueconvertedtype=DECIMAL, valuescale=2, valueprecision=5"`
-		ListWithDecimals []int32          `parquet:"name=list_with_decimals, type=LIST, valuetype=INT32, valueconvertedtype=DECIMAL, valuescale=2, valueprecision=5"`
-	}
-	nestedDecimalSchemaHandler, err := schema.NewSchemaHandlerFromStruct(new(NestedDecimalStruct))
-	require.NoError(t, err)
-
-	ptrValue := int32(1234)
-
-	tests := []struct {
-		name          string
-		input         any
-		schemaHandler *schema.SchemaHandler
-		expected      any
-	}{
-		{
-			name: "decimal_struct_with_various_types",
-			input: DecimalStruct{
-				Decimal1: 12345,
-				Decimal2: 98765432,
-				Decimal3: types.StrIntToBinary("1000000000000000", "BigEndian", 0, true),
-				Decimal4: types.StrIntToBinary("9876543210", "BigEndian", 12, true),
-				Decimal5: -4444,
-				Name:     "TestUser",
-				Age:      25,
-			},
-			schemaHandler: decimalSchemaHandler,
-			expected: map[string]any{
-				"Decimal1": float64(123.45),
-				"Decimal2": float64(98765.432),
-				"Decimal3": float64(100000000000.0000), // BYTE_ARRAY now also returns float64
-				"Decimal4": float64(98765432.10),       // FIXED_LEN_BYTE_ARRAY now also returns float64
-				"Decimal5": float64(-44.44),
-				"Name":     "TestUser",
-				"Age":      int32(25),
-			},
-		},
-		{
-			name: "nested_structures",
-			input: ContainerStruct{
-				Nested: NestedStruct{
-					Value: 9876,
-				},
-				SliceVal: []int32{100, 200, 300},
-				MapVal: map[string]int32{
-					"key1": 111,
-					"key2": 222,
-				},
-				PtrVal: &ptrValue,
-			},
-			schemaHandler: containerSchemaHandler,
-			expected: map[string]any{
-				"Nested":   map[string]any{"Value": float64(98.76)},
-				"SliceVal": []any{int32(100), int32(200), int32(300)},
-				"MapVal":   map[string]any{"key1": int32(111), "key2": int32(222)},
-				"PtrVal":   float64(123.4),
-			},
-		},
-		{
-			name:          "nil_value",
-			input:         nil,
-			schemaHandler: simpleSchemaHandler,
-			expected:      nil,
-		},
-		{
-			name: "slice_of_structs",
-			input: []SimpleStruct{
-				{Name: "test1"},
-				{Name: "test2"},
-			},
-			schemaHandler: simpleSchemaHandler,
-			expected: []any{
-				map[string]any{"Name": "test1"},
-				map[string]any{"Name": "test2"},
-			},
-		},
-		{
-			name: "map_with_string_keys",
-			input: map[string]SimpleStruct{
-				"key1": {Name: "value1"},
-				"key2": {Name: "value2"},
-			},
-			schemaHandler: simpleSchemaHandler,
-			expected: map[string]any{
-				"key1": map[string]any{"Name": "value1"},
-				"key2": map[string]any{"Name": "value2"},
-			},
-		},
-		{
-			name: "map_with_int_keys",
-			input: map[int]SimpleStruct{
-				5: {Name: "value5"},
-				7: {Name: "value7"},
-			},
-			schemaHandler: simpleSchemaHandler,
-			expected: map[string]any{
-				"5": map[string]any{"Name": "value5"},
-				"7": map[string]any{"Name": "value7"},
-			},
-		},
-		{
-			name:          "primitive_int",
-			input:         int32(42),
-			schemaHandler: simpleSchemaHandler,
-			expected:      int32(42),
-		},
-		{
-			name:          "primitive_string",
-			input:         "hello",
-			schemaHandler: simpleSchemaHandler,
-			expected:      "hello",
-		},
-		{
-			name:          "interface_types",
-			input:         any(SimpleStruct{Name: "interface_test"}),
-			schemaHandler: simpleSchemaHandler,
-			expected:      map[string]any{"Name": "interface_test"},
-		},
-		{
-			name:          "nil_interface",
-			input:         any(nil),
-			schemaHandler: simpleSchemaHandler,
-			expected:      nil,
-		},
-		{
-			name:          "pointer_to_struct_through_interface",
-			input:         any(&SimpleStruct{Name: "ptr_interface_test"}),
-			schemaHandler: simpleSchemaHandler,
-			expected:      map[string]any{"Name": "ptr_interface_test"},
-		},
-		{
-			name: "nested_decimals",
-			input: NestedDecimalStruct{
-				MapWithDecimals: map[string]int32{
-					"price1": 12345,
-					"price2": -6789,
-				},
-				ListWithDecimals: []int32{1111, 2222, -3333},
-			},
-			schemaHandler: nestedDecimalSchemaHandler,
-			expected: map[string]any{
-				"MapWithDecimals": map[string]any{
-					"price1": float64(123.45),
-					"price2": float64(-67.89),
-				},
-				"ListWithDecimals": []any{float64(11.11), float64(22.22), float64(-33.33)},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result, err := ConvertToJSONFriendly(tt.input, tt.schemaHandler)
-			require.NoError(t, err)
-			require.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestGetFieldNameFromTag(t *testing.T) {
-	tests := []struct {
-		name     string
-		field    reflect.StructField
-		expected string
-	}{
-		{
-			name: "json_tag_with_name",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:"json_name"`,
-			},
-			expected: "json_name",
-		},
-		{
-			name: "json_tag_with_options",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:"json_name,omitempty"`,
-			},
-			expected: "json_name",
-		},
-		{
-			name: "json_tag_with_multiple_options",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:"json_name,omitempty,string"`,
-			},
-			expected: "json_name",
-		},
-		{
-			name: "json_tag_dash_means_skip",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:"-"`,
-			},
-			expected: "FieldName", // Falls back to struct field name
-		},
-		{
-			name: "json_tag_empty_name",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:",omitempty"`,
-			},
-			expected: "FieldName", // Falls back to struct field name
-		},
-		{
-			name: "no_json_tag",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `parquet:"name=field_name"`,
-			},
-			expected: "FieldName", // Falls back to struct field name
-		},
-		{
-			name: "empty_tag",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  "",
-			},
-			expected: "FieldName", // Falls back to struct field name
-		},
-		{
-			name: "json_tag_empty",
-			field: reflect.StructField{
-				Name: "FieldName",
-				Tag:  `json:""`,
-			},
-			expected: "FieldName", // Falls back to struct field name
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := getFieldNameFromTag(tt.field)
-			require.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestConvertValueToJSONFriendlyWithContext(t *testing.T) {
-	type TestStruct struct {
-		Name  string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-		Value int32  `parquet:"name=value, type=INT32, convertedtype=DECIMAL, scale=2, precision=9"`
-	}
-
-	schemaHandler, err := schema.NewSchemaHandlerFromStruct(new(TestStruct))
-	require.NoError(t, err)
-
-	ctx := &jsonConverter{}
-
-	tests := []struct {
-		name          string
-		input         any
-		pathPrefix    string
-		expected      any
-		expectError   bool
-		useInvalidVal bool
-	}{
-		{
-			name:          "invalid_reflect_value",
-			input:         nil, // Will be ignored due to useInvalidVal
-			pathPrefix:    "",
-			expected:      nil,
-			expectError:   false,
-			useInvalidVal: true,
-		},
-		{
-			name:        "nil_interface",
-			input:       (interface{})(nil),
-			pathPrefix:  "",
-			expected:    nil,
-			expectError: false,
-		},
-		{
-			name:        "non_nil_interface",
-			input:       interface{}("test_string"),
-			pathPrefix:  "",
-			expected:    "test_string",
-			expectError: false,
-		},
-		{
-			name:        "nil_pointer",
-			input:       (*string)(nil),
-			pathPrefix:  "",
-			expected:    nil,
-			expectError: false,
-		},
-		{
-			name: "non_nil_pointer",
-			input: func() *string {
-				s := "test_string"
-				return &s
-			}(),
-			pathPrefix:  "",
-			expected:    "test_string",
-			expectError: false,
-		},
-		{
-			name:        "slice_conversion",
-			input:       []string{"a", "b", "c"},
-			pathPrefix:  "",
-			expected:    []any{"a", "b", "c"},
-			expectError: false,
-		},
-		{
-			name: "map_conversion",
-			input: map[string]int{
-				"key1": 1,
-				"key2": 2,
-			},
-			pathPrefix: "",
-			expected: map[string]any{
-				"key1": 1,
-				"key2": 2,
-			},
-			expectError: false,
-		},
-		{
-			name: "struct_conversion",
-			input: TestStruct{
-				Name:  "test",
-				Value: 1234,
-			},
-			pathPrefix: "",
-			expected: map[string]any{
-				"Name":  "test",
-				"Value": float64(12.34), // Converted due to decimal type
-			},
-			expectError: false,
-		},
-		{
-			name:        "primitive_conversion",
-			input:       int32(42),
-			pathPrefix:  "",
-			expected:    int32(42),
-			expectError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var val reflect.Value
-			if tt.useInvalidVal {
-				// Use invalid reflect value directly
-				val = reflect.Value{}
-			} else {
-				val = reflect.ValueOf(tt.input)
-			}
-
-			result, err := convertValueToJSONFriendlyWithContext(val, schemaHandler, tt.pathPrefix, ctx)
-
-			if tt.expectError {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tt.expected, result)
-			}
-		})
-	}
-}
-
-func TestConvertToJSONFriendly_NonDefaultRootName(t *testing.T) {
-	// Test the fix for hardcoded "Parquet_go_root" assumption
-	// Create a schema with a custom root name to verify the fix works
-
-	type TestStruct struct {
-		Name  string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-		Value int32  `parquet:"name=value, type=INT32, convertedtype=DECIMAL, scale=2, precision=9"`
-	}
-
-	// Create schema with custom root name
-	customRootSchema, err := schema.NewSchemaHandlerFromStruct(new(TestStruct))
-	require.NoError(t, err)
-
-	// Change the root name to test non-default scenario
-	customRootSchema.Infos[0].InName = "Custom_Root"
-	customRootSchema.Infos[0].ExName = "custom_root"
-
-	// Update the maps to reflect the new root name
-	oldMapIndex := make(map[string]int32)
-	for k, v := range customRootSchema.MapIndex {
-		oldMapIndex[k] = v
-	}
-
-	customRootSchema.MapIndex = make(map[string]int32)
-	for k, v := range oldMapIndex {
-		newKey := strings.Replace(k, common.ParGoRootInName, "Custom_Root", 1)
-		customRootSchema.MapIndex[newKey] = v
-	}
-
-	for k, v := range customRootSchema.IndexMap {
-		newValue := strings.Replace(v, common.ParGoRootInName, "Custom_Root", 1)
-		customRootSchema.IndexMap[k] = newValue
-	}
-
-	require.Equal(t, "Custom_Root", customRootSchema.GetRootInName())
-
-	// Test ConvertToJSONFriendly with custom root name
-	testData := TestStruct{
-		Name:  "TestUser",
-		Value: 12345,
-	}
-
-	result, err := ConvertToJSONFriendly(testData, customRootSchema)
-	require.NoError(t, err)
-
-	expected := map[string]any{
-		"Name":  "TestUser",
-		"Value": float64(123.45), // Converted due to decimal type
-	}
-	require.Equal(t, expected, result)
-}
-
-func TestConvertValueToJSONFriendlyWithContext_NilCases(t *testing.T) {
-	schemaHandler, err := schema.NewSchemaHandlerFromStruct(new(struct{}))
-	require.NoError(t, err)
-
-	ctx := &jsonConverter{}
-
-	tests := []struct {
-		name        string
-		setupValue  func() reflect.Value
-		expected    any
-		expectError bool
-	}{
-		{
-			name: "nil_interface_value",
-			setupValue: func() reflect.Value {
-				var nilInterface interface{} = nil
-				return reflect.ValueOf(&nilInterface).Elem()
-			},
-			expected:    nil,
-			expectError: false,
-		},
-		{
-			name: "nil_pointer_value",
-			setupValue: func() reflect.Value {
-				var nilPtr *string = nil
-				return reflect.ValueOf(nilPtr)
-			},
-			expected:    nil,
-			expectError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			val := tt.setupValue()
-			result, err := convertValueToJSONFriendlyWithContext(val, schemaHandler, "", ctx)
-
-			if tt.expectError {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tt.expected, result)
-			}
-		})
-	}
 }
 
 func TestUnmarshal_ErrorHandling_ReflectionSafety(t *testing.T) {
@@ -1099,444 +561,6 @@ func TestOldStyleList(t *testing.T) {
 }
 
 // Tests for shredded variant reconstruction functions
-
-func TestNewShreddedVariantReconstructor(t *testing.T) {
-	// Create schema with variant
-	variantInfo := &schema.VariantSchemaInfo{
-		MetadataIdx:    2,
-		ValueIdx:       3,
-		TypedValueIdxs: []int32{4},
-		IsShredded:     true,
-	}
-
-	sh := &schema.SchemaHandler{
-		IndexMap: map[int32]string{
-			2: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata",
-			3: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value",
-			4: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value",
-		},
-	}
-
-	metadataTable := &layout.Table{
-		Path:             []string{"Root", "Var", "Metadata"},
-		Values:           []any{[]byte{0x01, 0x00, 0x00}},
-		RepetitionLevels: []int32{0},
-		DefinitionLevels: []int32{1},
-	}
-	valueTable := &layout.Table{
-		Path:             []string{"Root", "Var", "Value"},
-		Values:           []any{[]byte{0x00}},
-		RepetitionLevels: []int32{0},
-		DefinitionLevels: []int32{2},
-	}
-	typedValueTable := &layout.Table{
-		Path:             []string{"Root", "Var", "Typed_value"},
-		Values:           []any{int32(42)},
-		RepetitionLevels: []int32{0},
-		DefinitionLevels: []int32{2},
-	}
-
-	tableMap := &map[string]*layout.Table{
-		"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata":    metadataTable,
-		"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":       valueTable,
-		"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value": typedValueTable,
-	}
-
-	t.Run("all_tables_present", func(t *testing.T) {
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			variantInfo,
-			tableMap,
-			sh,
-		)
-
-		require.NotNil(t, r)
-		require.Equal(t, "Root"+common.ParGoPathDelimiter+"Var", r.Path)
-		require.NotNil(t, r.MetadataTable)
-		require.NotNil(t, r.ValueTable)
-		require.Len(t, r.TypedValueTables, 1)
-	})
-
-	t.Run("no_value_table", func(t *testing.T) {
-		infoNoValue := &schema.VariantSchemaInfo{
-			MetadataIdx:    2,
-			ValueIdx:       -1, // No value column
-			TypedValueIdxs: []int32{4},
-			IsShredded:     true,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			infoNoValue,
-			tableMap,
-			sh,
-		)
-
-		require.NotNil(t, r)
-		require.NotNil(t, r.MetadataTable)
-		require.Nil(t, r.ValueTable)
-		require.Len(t, r.TypedValueTables, 1)
-	})
-
-	t.Run("missing_tables_in_map", func(t *testing.T) {
-		emptyTableMap := &map[string]*layout.Table{}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			variantInfo,
-			emptyTableMap,
-			sh,
-		)
-
-		require.NotNil(t, r)
-		require.Nil(t, r.MetadataTable)
-		require.Nil(t, r.ValueTable)
-		require.Empty(t, r.TypedValueTables)
-	})
-}
-
-func TestShreddedVariantReconstructor_GetValueAtRow(t *testing.T) {
-	sh := &schema.SchemaHandler{
-		MapIndex: map[string]int32{
-			"Root": 0,
-			"Root" + common.ParGoPathDelimiter + "Var":                                          1,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 2,
-		},
-		SchemaElements: []*parquet.SchemaElement{
-			{Name: "Root"},
-			{Name: "Var", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_OPTIONAL)},
-			{Name: "Metadata", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_REQUIRED)},
-		},
-	}
-
-	r := &ShreddedVariantReconstructor{
-		SchemaHandler: sh,
-	}
-
-	t.Run("nil_table", func(t *testing.T) {
-		result, err := r.getValueAtRow(nil, 0, nil, nil)
-		require.NoError(t, err)
-		require.Nil(t, result)
-	})
-
-	t.Run("missing_table_in_maps", func(t *testing.T) {
-		table := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		tableBgn := map[string]int{}
-		tableEnd := map[string]int{}
-
-		result, err := r.getValueAtRow(table, 0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Nil(t, result)
-	})
-
-	t.Run("negative_bgn", func(t *testing.T) {
-		table := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		tablePath := common.PathToStr(table.Path)
-		tableBgn := map[string]int{tablePath: -1}
-		tableEnd := map[string]int{tablePath: 1}
-
-		result, err := r.getValueAtRow(table, 0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Nil(t, result)
-	})
-
-	t.Run("valid_row_retrieval", func(t *testing.T) {
-		table := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01, 0x00, 0x00}, []byte{0x02, 0x00, 0x00}},
-			RepetitionLevels: []int32{0, 0},
-			DefinitionLevels: []int32{1, 1},
-		}
-		tablePath := common.PathToStr(table.Path)
-		tableBgn := map[string]int{tablePath: 0}
-		tableEnd := map[string]int{tablePath: 2}
-
-		result, err := r.getValueAtRow(table, 0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Equal(t, []byte{0x01, 0x00, 0x00}, result)
-
-		result, err = r.getValueAtRow(table, 1, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Equal(t, []byte{0x02, 0x00, 0x00}, result)
-	})
-
-	t.Run("row_not_found", func(t *testing.T) {
-		table := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01, 0x00, 0x00}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		tablePath := common.PathToStr(table.Path)
-		tableBgn := map[string]int{tablePath: 0}
-		tableEnd := map[string]int{tablePath: 1}
-
-		result, err := r.getValueAtRow(table, 5, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Nil(t, result)
-	})
-
-	t.Run("null_value_low_definition_level", func(t *testing.T) {
-		// When definition level is less than max, value is null
-		table := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{nil},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{0}, // Lower than max
-		}
-		tablePath := common.PathToStr(table.Path)
-		tableBgn := map[string]int{tablePath: 0}
-		tableEnd := map[string]int{tablePath: 1}
-
-		result, err := r.getValueAtRow(table, 0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Nil(t, result)
-	})
-}
-
-func TestShreddedVariantReconstructor_Reconstruct(t *testing.T) {
-	sh := &schema.SchemaHandler{
-		MapIndex: map[string]int32{
-			"Root": 0,
-			"Root" + common.ParGoPathDelimiter + "Var":                                             1,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata":    2,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":       3,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value": 4,
-		},
-		IndexMap: map[int32]string{
-			0: "Root",
-			1: "Root" + common.ParGoPathDelimiter + "Var",
-			2: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata",
-			3: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value",
-			4: "Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value",
-		},
-		SchemaElements: []*parquet.SchemaElement{
-			{Name: "Root"},
-			{Name: "Var", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_OPTIONAL)},
-			{Name: "Metadata", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_REQUIRED)},
-			{Name: "Value", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_OPTIONAL)},
-			{Name: "Typed_value", RepetitionType: common.ToPtr(parquet.FieldRepetitionType_OPTIONAL)},
-		},
-	}
-
-	t.Run("metadata_as_bytes", func(t *testing.T) {
-		metadataTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01, 0x00, 0x00}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		valueTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Value"},
-			Values:           []any{[]byte{0x00}}, // null variant value
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{2},
-		}
-
-		tableMap := &map[string]*layout.Table{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    valueTable,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			&schema.VariantSchemaInfo{
-				MetadataIdx:    2,
-				ValueIdx:       3,
-				TypedValueIdxs: nil,
-				IsShredded:     false,
-			},
-			tableMap,
-			sh,
-		)
-
-		tableBgn := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 0,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    0,
-		}
-		tableEnd := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 1,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    1,
-		}
-
-		variant, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.NotNil(t, variant.Metadata)
-		require.Equal(t, []byte{0x01, 0x00, 0x00}, variant.Metadata)
-	})
-
-	t.Run("metadata_as_string", func(t *testing.T) {
-		metadataTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{string([]byte{0x01, 0x00, 0x00})},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-
-		tableMap := &map[string]*layout.Table{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			&schema.VariantSchemaInfo{
-				MetadataIdx:    2,
-				ValueIdx:       -1,
-				TypedValueIdxs: nil,
-				IsShredded:     false,
-			},
-			tableMap,
-			sh,
-		)
-
-		tableBgn := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 0,
-		}
-		tableEnd := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 1,
-		}
-
-		variant, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.NotNil(t, variant.Metadata)
-	})
-
-	t.Run("metadata_unexpected_type", func(t *testing.T) {
-		metadataTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{int32(123)}, // Unexpected type
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-
-		tableMap := &map[string]*layout.Table{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			&schema.VariantSchemaInfo{
-				MetadataIdx:    2,
-				ValueIdx:       -1,
-				TypedValueIdxs: nil,
-				IsShredded:     false,
-			},
-			tableMap,
-			sh,
-		)
-
-		tableBgn := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 0,
-		}
-		tableEnd := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 1,
-		}
-
-		_, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.Error(t, err)
-		require.Contains(t, err.Error(), "unexpected metadata type")
-	})
-
-	t.Run("value_as_string", func(t *testing.T) {
-		metadataTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01, 0x00, 0x00}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		valueTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Value"},
-			Values:           []any{string([]byte{0x00})},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{2},
-		}
-
-		tableMap := &map[string]*layout.Table{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    valueTable,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			&schema.VariantSchemaInfo{
-				MetadataIdx:    2,
-				ValueIdx:       3,
-				TypedValueIdxs: nil,
-				IsShredded:     false,
-			},
-			tableMap,
-			sh,
-		)
-
-		tableBgn := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 0,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    0,
-		}
-		tableEnd := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": 1,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    1,
-		}
-
-		variant, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.NotNil(t, variant.Value)
-	})
-
-	t.Run("with_typed_value", func(t *testing.T) {
-		metadataTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Metadata"},
-			Values:           []any{[]byte{0x01, 0x00, 0x00}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		typedValueTable := &layout.Table{
-			Path:             []string{"Root", "Var", "Typed_value"},
-			Values:           []any{int32(12345)},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{2},
-		}
-
-		tableMap := &map[string]*layout.Table{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata":    metadataTable,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value": typedValueTable,
-		}
-
-		r := NewShreddedVariantReconstructor(
-			"Root"+common.ParGoPathDelimiter+"Var",
-			&schema.VariantSchemaInfo{
-				MetadataIdx:    2,
-				ValueIdx:       -1,
-				TypedValueIdxs: []int32{4},
-				IsShredded:     true,
-			},
-			tableMap,
-			sh,
-		)
-
-		tableBgn := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata":    0,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value": 0,
-		}
-		tableEnd := map[string]int{
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata":    1,
-			"Root" + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Typed_value": 1,
-		}
-
-		variant, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.NotNil(t, variant.Value)
-	})
-}
 
 func TestSetVariantValue(t *testing.T) {
 	t.Run("simple_struct_with_variant", func(t *testing.T) {
@@ -1954,583 +978,4 @@ func TestUnmarshal_ErrorHandling_TypeConversion(t *testing.T) {
 	// We expect an error here, not a panic.
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "cannot convert value of type")
-}
-
-func TestShreddedVariantReconstructor_Recursive(t *testing.T) {
-	// Root.Var (VARIANT) shredded into:
-	// Root.Var.Metadata
-	// Root.Var.Typed_value.a.Metadata
-	// Root.Var.Typed_value.a.Value
-
-	// metadata: dict={0=>a}
-	metadata := types.EncodeVariantMetadata([]string{"A"})
-
-	// Inner variant for "a": value=123 (INT8)
-	innerVariant := types.Variant{
-		Metadata: metadata,
-		Value:    types.EncodeVariantInt8(123),
-	}
-
-	mKey := common.PathToStr([]string{common.ParGoRootInName, "Var", "Metadata"})
-	vKey := common.PathToStr([]string{common.ParGoRootInName, "Var", "Value"})
-	imKey := common.PathToStr([]string{common.ParGoRootInName, "Var", "TypedValue", "A", "Metadata"})
-	ivKey := common.PathToStr([]string{common.ParGoRootInName, "Var", "TypedValue", "A", "Value"})
-
-	tableMap := map[string]*layout.Table{
-		mKey: {
-			Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-			Values:           []any{metadata},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		},
-		vKey: {
-			Path:             []string{common.ParGoRootInName, "Var", "Value"},
-			Values:           []any{nil},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{0},
-		},
-		imKey: {
-			Path:             []string{common.ParGoRootInName, "Var", "TypedValue", "A", "Metadata"},
-			Values:           []any{metadata},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		},
-		ivKey: {
-			Path:             []string{common.ParGoRootInName, "Var", "TypedValue", "A", "Value"},
-			Values:           []any{innerVariant.Value},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		},
-	}
-
-	type Inner struct {
-		Metadata []byte `parquet:"name=Metadata, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-		Value    []byte `parquet:"name=Value, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-	}
-	type TypedValue struct {
-		A Inner `parquet:"name=A, repetitiontype=REQUIRED"`
-	}
-	type VarGroup struct {
-		Metadata   []byte     `parquet:"name=Metadata, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-		Value      []byte     `parquet:"name=Value, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-		TypedValue TypedValue `parquet:"name=TypedValue, repetitiontype=REQUIRED"`
-	}
-
-	sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-		Var VarGroup `parquet:"name=Var, type=VARIANT, repetitiontype=REQUIRED"`
-	}))
-
-	var info *schema.VariantSchemaInfo
-	var path string
-	for p, i := range sh.VariantSchemas {
-		path = p
-		info = i
-		break
-	}
-
-	r := NewShreddedVariantReconstructor(path, info, &tableMap, sh)
-
-	tableBgn := map[string]int{mKey: 0, vKey: 0, imKey: 0, ivKey: 0}
-	tableEnd := map[string]int{mKey: 1, vKey: 1, imKey: 1, ivKey: 1}
-
-	v, err := r.Reconstruct(0, tableBgn, tableEnd)
-	require.NoError(t, err)
-
-	decoded, err := types.ConvertVariantValue(v)
-	require.NoError(t, err)
-
-	expected := map[string]any{"A": int8(123)}
-	require.Equal(t, expected, decoded)
-}
-
-func TestShreddedVariantReconstructor_Reconstruct_Coverage(t *testing.T) {
-	t.Run("reconstruct_value_error_propagation", func(t *testing.T) {
-		metadata := types.EncodeVariantMetadata([]string{"a"})
-		metadataTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-			Values:           []any{metadata},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-
-		// Incompatible value to trigger error in reconstructValue or getValueAtRow
-		valueTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "Value"},
-			Values:           []any{struct{ Bad bool }{Bad: true}},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-
-		tableMap := map[string]*layout.Table{
-			common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-			common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    valueTable,
-		}
-
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var any `parquet:"name=Var, type=VARIANT"`
-		}))
-
-		var info *schema.VariantSchemaInfo
-		var path string
-		for p, i := range sh.VariantSchemas {
-			path = p
-			info = i
-			break
-		}
-
-		r := NewShreddedVariantReconstructor(path, info, &tableMap, sh)
-
-		mKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata"
-		vKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value"
-
-		tableBgn := map[string]int{mKey: 0, vKey: 0}
-		tableEnd := map[string]int{mKey: 1, vKey: 1}
-
-		_, err := r.Reconstruct(0, tableBgn, tableEnd)
-		// Should error because we passed a struct as BYTE_ARRAY value
-		require.Error(t, err)
-		require.Contains(t, err.Error(), "unexpected value type")
-	})
-
-	t.Run("reconstruct_returns_existing_variant", func(t *testing.T) {
-		metadata := types.EncodeVariantMetadata([]string{"a"})
-		expectedVariant := types.Variant{
-			Metadata: metadata,
-			Value:    types.EncodeVariantInt8(123),
-		}
-
-		// tableMap containing a Variant struct directly (if that's possible in some paths)
-		// Actually reconstructValue returns Variant if it finds Metadata/Value children
-		metadataTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-			Values:           []any{metadata},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		valueTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "Value"},
-			Values:           []any{expectedVariant.Value},
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-
-		tableMap := map[string]*layout.Table{
-			common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata": metadataTable,
-			common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value":    valueTable,
-		}
-
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var any `parquet:"name=Var, type=VARIANT"`
-		}))
-
-		var info *schema.VariantSchemaInfo
-		var path string
-		for p, i := range sh.VariantSchemas {
-			path = p
-			info = i
-			break
-		}
-
-		r := NewShreddedVariantReconstructor(path, info, &tableMap, sh)
-
-		mKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata"
-		vKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value"
-
-		tableBgn := map[string]int{mKey: 0, vKey: 0}
-		tableEnd := map[string]int{mKey: 1, vKey: 1}
-
-		v, err := r.Reconstruct(0, tableBgn, tableEnd)
-		require.NoError(t, err)
-		require.Equal(t, expectedVariant, v)
-	})
-}
-
-func TestShreddedVariantReconstructor_reconstructValue_Coverage(t *testing.T) {
-	t.Run("nested_list_reconstruction", func(t *testing.T) {
-		// Mock a LIST of Variants
-		// Root.Var (LIST)
-		//   List
-		//     Element (VARIANT)
-		//       Metadata
-		//       Value
-
-		metadata := types.EncodeVariantMetadata([]string{})
-		val1 := types.EncodeVariantInt8(1)
-		val2 := types.EncodeVariantInt8(2)
-
-		mKeyInt := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "List" + common.ParGoPathDelimiter + "Element" + common.ParGoPathDelimiter + "Metadata"
-		vKeyInt := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "List" + common.ParGoPathDelimiter + "Element" + common.ParGoPathDelimiter + "Value"
-
-		metadataTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "List", "Element", "Metadata"},
-			Values:           []any{metadata, metadata},
-			RepetitionLevels: []int32{0, 1},
-			DefinitionLevels: []int32{1, 1},
-		}
-		valueTable := &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "List", "Element", "Value"},
-			Values:           []any{val1, val2},
-			RepetitionLevels: []int32{0, 1},
-			DefinitionLevels: []int32{1, 1},
-		}
-
-		tableMap := map[string]*layout.Table{
-			mKeyInt: metadataTable,
-			vKeyInt: valueTable,
-		}
-
-		type ListElement struct {
-			Element any `parquet:"name=element, type=VARIANT"`
-		}
-		type ListType struct {
-			List []ListElement `parquet:"name=list, repetitiontype=REPEATED"`
-		}
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var ListType `parquet:"name=Var, type=LIST"`
-		}))
-
-		// For LIST, we need to make sure MaxRepetitionLevel > 0 for children
-		// But here we are calling reconstructValue on the LIST group itself.
-		// reconstructValue handles LIST by going to List/Element.
-
-		// Manually create reconstructor since it's not a top-level Variant group
-		r := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap,
-			SchemaHandler: sh,
-		}
-
-		tableBgn := map[string]int{mKeyInt: 0, vKeyInt: 0}
-		tableEnd := map[string]int{mKeyInt: 2, vKeyInt: 2}
-
-		res, err := r.reconstructValue(r.Path, 0, tableBgn, tableEnd, nil)
-		require.NoError(t, err)
-
-		// For LIST, reconstructValue should return a slice
-		slice, ok := res.([]any)
-		require.True(t, ok, "Expected slice, got %T", res)
-		require.Len(t, slice, 2)
-	})
-
-	t.Run("map_reconstruction_zip_slices", func(t *testing.T) {
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Obj []struct {
-				A int32  `parquet:"name=a, type=INT32"`
-				B string `parquet:"name=b, type=BYTE_ARRAY, convertedtype=UTF8"`
-			} `parquet:"name=Obj, repetitiontype=REPEATED"`
-		}))
-
-		// Internal paths for repeated struct fields
-		aPath := []string{common.ParGoRootInName, "Obj", "A"}
-		bPath := []string{common.ParGoRootInName, "Obj", "B"}
-		aKey := common.PathToStr(aPath)
-		bKey := common.PathToStr(bPath)
-
-		tableMap := map[string]*layout.Table{
-			aKey: {
-				Path:             aPath,
-				Values:           []any{int32(1), int32(2)},
-				RepetitionLevels: []int32{0, 1},
-				DefinitionLevels: []int32{1, 1},
-			},
-			bKey: {
-				Path:             bPath,
-				Values:           []any{"x", "y"},
-				RepetitionLevels: []int32{0, 1},
-				DefinitionLevels: []int32{1, 1},
-			},
-		}
-
-		r := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Obj",
-			tableMap:      &tableMap,
-			SchemaHandler: sh,
-		}
-
-		tableBgn := map[string]int{aKey: 0, bKey: 0}
-		tableEnd := map[string]int{aKey: 2, bKey: 2}
-
-		res, err := r.reconstructValue(r.Path, 0, tableBgn, tableEnd, nil)
-		require.NoError(t, err)
-
-		slice, ok := res.([]any)
-		require.True(t, ok, "Expected slice, got %T", res)
-		require.Len(t, slice, 2)
-
-		m1 := slice[0].(map[string]any)
-		require.Equal(t, int32(1), m1["a"])
-		require.Equal(t, "x", m1["b"])
-
-		m2 := slice[1].(map[string]any)
-		require.Equal(t, int32(2), m2["a"])
-		require.Equal(t, "y", m2["b"])
-	})
-
-	t.Run("metadata_as_string_and_variant_reconstruction", func(t *testing.T) {
-		// 1. Test regular object reconstruction (non-variant, non-list)
-		sh1, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var struct {
-				A int32 `parquet:"name=A, type=INT32"`
-			} `parquet:"name=Var"`
-		}))
-
-		aKey1 := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "A"
-		tableMap1 := map[string]*layout.Table{
-			aKey1: {
-				Path:             []string{common.ParGoRootInName, "Var", "A"},
-				Values:           []any{int32(42)},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-		}
-
-		r1 := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap1,
-			SchemaHandler: sh1,
-		}
-
-		tableBgn1 := map[string]int{aKey1: 0}
-		tableEnd1 := map[string]int{aKey1: 1}
-
-		res1, err := r1.reconstructValue(r1.Path, 0, tableBgn1, tableEnd1, nil)
-		require.NoError(t, err)
-		m1 := res1.(map[string]any)
-		require.Equal(t, int32(42), m1["A"])
-
-		// 2. Test metadata as string in a VARIANT group
-		type VarGroup struct {
-			Metadata []byte `parquet:"name=Metadata, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-			Value    []byte `parquet:"name=Value, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-		}
-		sh2, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var VarGroup `parquet:"name=Var, type=VARIANT, repetitiontype=REQUIRED"`
-		}))
-		mKey2 := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata"
-		vKey2 := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value"
-		tableMap2 := map[string]*layout.Table{
-			mKey2: {
-				Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-				Values:           []any{string([]byte{0x01, 0x00, 0x00})},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-			vKey2: {
-				Path:             []string{common.ParGoRootInName, "Var", "Value"},
-				Values:           []any{types.EncodeVariantInt32(123)},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-		}
-		r2 := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap2,
-			SchemaHandler: sh2,
-		}
-		tableBgn2 := map[string]int{mKey2: 0, vKey2: 0}
-		tableEnd2 := map[string]int{mKey2: 1, vKey2: 1}
-
-		res2, err := r2.reconstructValue(r2.Path, 0, tableBgn2, tableEnd2, nil)
-		require.NoError(t, err)
-		v2, ok := res2.(types.Variant)
-		require.True(t, ok, "Expected types.Variant, got %T", res2)
-		require.Equal(t, []byte{0x01, 0x00, 0x00}, v2.Metadata)
-
-		// 3. Trigger ReconstructVariant error fallback
-		type VarGroupTyped struct {
-			Metadata   []byte `parquet:"name=Metadata, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-			Value      []byte `parquet:"name=Value, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-			TypedValue []byte `parquet:"name=TypedValue, type=BYTE_ARRAY, repetitiontype=REQUIRED"`
-		}
-		sh3, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var VarGroupTyped `parquet:"name=Var, type=VARIANT, repetitiontype=REQUIRED"`
-		}))
-		tKey3 := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "TypedValue"
-		tableMap2[tKey3] = &layout.Table{
-			Path:             []string{common.ParGoRootInName, "Var", "TypedValue"},
-			Values:           []any{func() {}}, // Incompatible type for variant encoding
-			RepetitionLevels: []int32{0},
-			DefinitionLevels: []int32{1},
-		}
-		tableBgn2[tKey3], tableEnd2[tKey3] = 0, 1
-
-		r3 := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap2,
-			SchemaHandler: sh3,
-		}
-
-		res3, err := r3.reconstructValue(r3.Path, 0, tableBgn2, tableEnd2, nil)
-		require.NoError(t, err)
-		require.Nil(t, res3)
-	})
-
-	t.Run("reconstruct_null_and_variant_direct", func(t *testing.T) {
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var int32 `parquet:"name=Var, type=VARIANT"`
-		}))
-		r := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &map[string]*layout.Table{},
-			SchemaHandler: sh,
-		}
-		// Null case
-		v, err := r.Reconstruct(0, map[string]int{}, map[string]int{})
-		require.NoError(t, err)
-		valNull, _ := types.ConvertVariantValue(v)
-		require.Nil(t, valNull)
-
-		// Variant direct case
-		mKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata"
-		vKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value"
-		tableMap := map[string]*layout.Table{
-			mKey: {
-				Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-				Values:           []any{[]byte{0x01, 0x00, 0x00}},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-			vKey: {
-				Path:             []string{common.ParGoRootInName, "Var", "Value"},
-				Values:           []any{types.EncodeVariantInt32(123)},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-		}
-		r.tableMap = &tableMap
-		v, err = r.Reconstruct(0, map[string]int{mKey: 0, vKey: 0}, map[string]int{mKey: 1, vKey: 1})
-		require.NoError(t, err)
-		valDirect, _ := types.ConvertVariantValue(v)
-		require.Equal(t, int32(123), valDirect)
-	})
-
-	t.Run("reconstruct_any_to_variant", func(t *testing.T) {
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var int32 `parquet:"name=Var, type=INT32"`
-		}))
-		tKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var"
-		tableMap := map[string]*layout.Table{
-			tKey: {
-				Path:             []string{common.ParGoRootInName, "Var"},
-				Values:           []any{int32(123)},
-				RepetitionLevels: []int32{0},
-				DefinitionLevels: []int32{1},
-			},
-		}
-		r := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap,
-			SchemaHandler: sh,
-		}
-		v, err := r.Reconstruct(0, map[string]int{tKey: 0}, map[string]int{tKey: 1})
-		require.NoError(t, err)
-		valAny, _ := types.ConvertVariantValue(v)
-		require.Equal(t, int32(123), valAny)
-	})
-
-	t.Run("value_as_string_and_is_repeated_from_children", func(t *testing.T) {
-		sh, _ := schema.NewSchemaHandlerFromStruct(new(struct {
-			Var []int32 `parquet:"name=Var, type=VARIANT, repetitiontype=REPEATED"`
-		}))
-		mKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Metadata"
-		vKey := common.ParGoRootInName + common.ParGoPathDelimiter + "Var" + common.ParGoPathDelimiter + "Value"
-
-		// Metadata and Value as strings, and repeated
-		tableMap := map[string]*layout.Table{
-			mKey: {
-				Path:             []string{common.ParGoRootInName, "Var", "Metadata"},
-				Values:           []any{string([]byte{0x01, 0x00, 0x00}), string([]byte{0x01, 0x00, 0x00})},
-				RepetitionLevels: []int32{0, 1},
-				DefinitionLevels: []int32{1, 1},
-			},
-			vKey: {
-				Path:             []string{common.ParGoRootInName, "Var", "Value"},
-				Values:           []any{string(types.EncodeVariantInt32(111)), string(types.EncodeVariantInt32(222))},
-				RepetitionLevels: []int32{0, 1},
-				DefinitionLevels: []int32{1, 1},
-			},
-		}
-
-		r := &ShreddedVariantReconstructor{
-			Path:          common.ParGoRootInName + common.ParGoPathDelimiter + "Var",
-			tableMap:      &tableMap,
-			SchemaHandler: sh,
-		}
-
-		tableBgn := map[string]int{mKey: 0, vKey: 0}
-		tableEnd := map[string]int{mKey: 2, vKey: 2}
-
-		res, err := r.reconstructValue(r.Path, 0, tableBgn, tableEnd, nil)
-		require.NoError(t, err)
-
-		slice, ok := res.([]any)
-		require.True(t, ok, "Expected slice due to repeated children, got %T", res)
-		require.Len(t, slice, 2)
-
-		v1 := slice[0].(types.Variant)
-		val1, _ := types.ConvertVariantValue(v1)
-		require.Equal(t, int32(111), val1)
-
-		v2 := slice[1].(types.Variant)
-		val2, _ := types.ConvertVariantValue(v2)
-		require.Equal(t, int32(222), val2)
-	})
-}
-
-func TestConvertToJSONFriendly_WithGeospatialConfig(t *testing.T) {
-	type GeoRow struct {
-		Geom string `parquet:"name=geom, type=BYTE_ARRAY, logicaltype=GEOMETRY"`
-	}
-	sh, err := schema.NewSchemaHandlerFromStruct(new(GeoRow))
-	require.NoError(t, err)
-
-	// WKB Point(100, 50) little-endian
-	wkbPoint := string([]byte{0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x59, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x49, 0x40})
-
-	tests := []struct {
-		name     string
-		opts     []JSONConvertOption
-		expected map[string]any
-	}{
-		{
-			name: "default_uses_hex_mode",
-			opts: nil,
-			expected: map[string]any{
-				"wkb_hex": "010100000000000000000059400000000000004940",
-				"crs":     "OGC:CRS84",
-			},
-		},
-		{
-			name: "custom_geojson_mode",
-			opts: []JSONConvertOption{WithGeospatialConfig(types.NewGeospatialConfig(
-				types.WithGeometryJSONMode(types.GeospatialModeGeoJSON),
-			))},
-			expected: map[string]any{
-				"type":       "Feature",
-				"geometry":   map[string]any{"type": "Point", "coordinates": []float64{100, 50}},
-				"properties": map[string]any{"crs": "OGC:CRS84"},
-			},
-		},
-		{
-			name: "custom_base64_mode",
-			opts: []JSONConvertOption{WithGeospatialConfig(types.NewGeospatialConfig(
-				types.WithGeometryJSONMode(types.GeospatialModeBase64),
-			))},
-			expected: map[string]any{
-				"wkb_b64": "AQEAAAAAAAAAAABZQAAAAAAAAElA",
-				"crs":     "OGC:CRS84",
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			input := GeoRow{Geom: wkbPoint}
-			result, err := ConvertToJSONFriendly(input, sh, tt.opts...)
-			require.NoError(t, err)
-			resultMap, ok := result.(map[string]any)
-			require.True(t, ok)
-			require.Equal(t, tt.expected, resultMap["Geom"])
-		})
-	}
 }
